Update debian/copyright
Read in NOTICE.  Upstream has reformatted things slightly, and
their copyright is now through 2013.

Add a common-licenses reference for the GPL-2+ reference in the
libev license.  (It did not include the full GPL text inline, just
a reference, but lintian complains if we don't provide the reference
with an explicit target.)