Add symbols files for the kadm5 libraries
The KADM5 API version number was increased for the 1.11 release but the
corresponding library sonames were not, so there is an incompatible
change in the behavior of all the kadm5 library routines.  As such,
we must indicate the behavior change ourself to get the correct
behavior from versioned library dependencies.