Add libkrad and krb5-otp
A new preauth module for the KDC and its support library.

The API of libkrad is not expected to be stable; its symbols are public
but we need a strict version dependency on the library.  Future changes
to the library will require close scrutiny as we will likely need to
adjust the versions in the symbols file with future changes.