Update symbols files
Synchronize with the output of dpkg-gensymbols.

This catches up with the kadm5 soname bump as well.

These do match the changes in upstream between 1.11 and 1.12.
Most removed symbols are internal symbols which were exposed due
to implementation details.  The implementation no longer requires
them to be exposed, so upstream does not expose them anymore.  There
are also some removed symbols in the kadm5 libraries, but those
are not considered stable interfaces by upstream (and we're making
a new package anyway).

I'm not sure why dpkg-gensyms decided to change its sorting (?) for
some of these symbols.