Add a versioned Breaks: to libkrb5support0
The library export symbol list for libkrb5support0 changed with the
import of upstream's 1.12 release.  In particular, some symbols were renamed
and are no longer provided by new versions of the library.  This means
that new versions of libkrb5support0 are not compatible with old versions
of libraries that consume those symbols (almost all the libraries produced
from src:krb5), since we normally only insert >= dependencies on
libkrb5support0.

The ABI of libkrb5support0 is not guaranteed to stable, to it would be
technically correct to require an = dependency on it from all those
consumer libraries, but in practice it is mostly stable, so we elect
to only require >= and put Breaks: in place where applicable.  This
is one such place.