Debian changes for runstatedir being /run
We have to do this via patches for now, since configure --runstatedir
is only in autoconf 2.70 which we don't have yet.

runstatedir is now used, as the default location of the socket for OTP
verification.  Since this is a change with potential to break
existing setups, announce it in krb5-otp.NEWS.