krb5-k5tls and krb5-otp depend on recent-ish libkrb5s
The libkrb5 has to be new enough to support the plugin interface
for which the given module is trying to register.  krb5-otp needs
a 1.12 variant, and krb5-k5tls needs a 1.13 variant.

Make these dependencies explicit, as dpkg-shlibdeps will not detect
them properly (PLUGIN_INTERFACE_* are macros and are also not public
symbols).